Converted REACT JS to REACT TS
