+++ index.html
@@ -1,27 +1,13 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- <link rel="icon" type="image/svg+xml" href="/favicon.ico" /> -->
-    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
-    <title>Preston Mayieka</title>
-    <meta
-      name="description"
-      content="This is the official Portfolio of Preston Mayieka"
-    />
-    <meta name="robots" content="follow, index" />
+    <title>Vite + React + TS</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-    <!-- Google tag (gtag.js) -->
-<script async src="https://www.googletagmanager.com/gtag/js?id=G-HSYH0BM0Q3"></script>
-<script>
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-  gtag('config', 'G-HSYH0BM0Q3');
-</script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
