+++ package.json
@@ -1,12 +1,12 @@
 {
-  "name": "react-portfolio-project",
+  "name": "ts",
   "private": true,
   "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
-    "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -20,21 +20,25 @@
     "react-helmet": "^6.1.0",
     "react-preloaders": "^3.0.3",
     "react-router-dom": "^6.22.0",
-    "react-scroll": "^1.8.9",
+    "react-scroll": "^1.9.0",
     "react-social-icons": "^5.15.0",
     "react-toastify": "^10.0.4",
     "typewriter-effect": "^2.20.1"
   },
   "devDependencies": {
-    "@types/react": "^18.0.37",
-    "@vitejs/plugin-react": "^4.0.0",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.38.0",
-    "eslint-plugin-react": "^7.32.2",
+    "@types/node": "^20.11.29",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
+    "@typescript-eslint/eslint-plugin": "^7.1.1",
+    "@typescript-eslint/parser": "^7.1.1",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.18",
+    "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.3.4",
-    "postcss": "^8.4.24",
-    "tailwindcss": "^3.3.2",
-    "vite": "^4.3.9"
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.36",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.1.6"
   }
 }
