+++ src/Pages/LandingPage.tsx
@@ -4,7 +4,7 @@ import Home from '../components/Home';
 import About from '../components/About';
 import Projects from '../components/Projects';
 import Blogs from '../components/Blogs';
-import Stats from '../components/Stats';
+// import Stats from '../components/Stats';
 import Contact from '../components/Contact';
 import Footer from '../components/Footer';
 import whatsappLogo from '/assets/icons/WhatsApp_icon.png';
@@ -12,8 +12,8 @@ import backToTopIcon from '/assets/icons/back-to-top.png';
 
 
 const LandingPage = () => {
-    const [showBackToTop, setShowBackToTop] = useState(false);
-    const [showWhatsApp, setShowWhatsApp] = useState(true);
+    const [showBackToTop, setShowBackToTop] = useState<boolean>(false);
+    const [showWhatsApp, setShowWhatsApp] = useState<boolean>(true);
 
     useEffect(() => {
         const handleScroll = () => {
