+++ src/components/SocialIconsContainer.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { SocialIcon } from 'react-social-icons'
 
 const SocialIcons = () => {
